backend/requirements.txt
alembic==1.11.1
aniso8601==9.0.1
attrs==23.1.0
blinker==1.6.2
brotlipy==0.7.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
clickclick==20.10.2
connexion==2.14.2
cryptography==40.0.2
Flask==2.2.5
flask-marshmallow==0.15.0
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
greenlet==2.0.2
idna==3.4
importlib-metadata==6.6.0
inflection==0.5.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.2
marshmallow==3.19.0
marshmallow-sqlalchemy==0.29.0
packaging==23.1
pip==23.1.2
psycopg2==2.9.6
pycparser==2.21
pycryptodome==3.18.0
pyOpenSSL==23.1.1
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
requests==2.29.0
setuptools==67.7.2
six==1.16.0
SQLAlchemy==2.0.14
swagger-ui-bundle==0.0.9
typing_extensions==4.5.0
urllib3==1.26.15
Werkzeug==2.2.3
wheel==0.40.0
zipp==3.15.0
